+++ R/recipes-step_smooth.R
@@ -283,8 +283,8 @@ bake.step_smooth <- function(object, new_data, ...) {
 #' @export
 print.step_smooth <-
     function(x, width = max(20, options()$width - 35), ...) {
-        cat("Smoother: Local Polynomial Regression Fitting (Loess) on ")
-        recipes::print_step(x$columns, x$terms, x$trained, width = width)
+        title <- "Smoother: Local Polynomial Regression Fitting (Loess) on "
+        recipes::print_step(x$columns, x$terms, x$trained, width = width, title = title)
         invisible(x)
     }