+++ R/recipes-step_ts_clean.R
@@ -188,8 +188,8 @@ bake.step_ts_clean <- function(object, new_data, ...) {
 
 #' @export
 print.step_ts_clean <- function(x, width = max(20, options()$width - 35), ...) {
-    cat("Time Series Outlier Cleaning on ", sep = "")
-    recipes::print_step(names(x$lambdas_trained), x$terms, x$trained, width = width)
+    title <- "Time Series Outlier Cleaning on "
+    recipes::print_step(names(x$lambdas_trained), x$terms, x$trained, width = width, title = title)
     invisible(x)
 }